requirements.txt
python-decouple==3.4
Python-Jenkins==1.7.0